+++ analysis_options.yaml
@@ -1,1 +1,1 @@
-#include: package:surf_lint_rules/analysis_options.yaml
+include: package:surf_lint_rules/analysis_options.yaml
