+++ lib/ui/res/app_routes.dart
@@ -1,3 +1,4 @@
+import 'package:dart_testing/domain/testing.dart';
 import 'package:dart_testing/ui/screens/homeScreen/home_screen.dart';
 import 'package:dart_testing/ui/screens/resultScreen/result_screen.dart';
 import 'package:dart_testing/ui/screens/testingScreen/testing_screen.dart';
@@ -5,20 +6,29 @@ import 'package:flutter/material.dart';
 
 /// основные маршруты приложения
 class AppRoutes {
+  AppRoutes._();
+
   /// перейти на экран тестирования
   static Future<Object?> goTestingScreen(BuildContext context) {
     return Navigator.of(context).pushReplacement(
       MaterialPageRoute(
-        builder: (context) => TestingScreen(),
+        builder: (context) => const TestingScreen(),
       ),
     );
   }
 
   /// перейти на экран с результатами
-  static Future<Object?> goResultScreen(BuildContext context) {
+  static Future<Object?> goResultScreen(
+    BuildContext context, {
+    required ResultTesting resultTesting,
+    required List<bool> results,
+  }) {
     return Navigator.of(context).pushReplacement(
       MaterialPageRoute(
-        builder: (context) => ResultScreen(),
+        builder: (context) => ResultScreen(
+          resultTesting: resultTesting,
+          results: results,
+        ),
       ),
     );
   }
@@ -27,10 +37,8 @@ class AppRoutes {
   static Future<Object?> goHomeScreen(BuildContext context) {
     return Navigator.of(context).pushReplacement(
       MaterialPageRoute(
-        builder: (context) => HomeScreen(),
+        builder: (context) => const HomeScreen(),
       ),
     );
   }
-
-  AppRoutes._();
 }
