--- lib/ui/widgets/custom_text_button.dart
@@ -1,33 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// Текстовая кнопка
-class CustomTextButton extends StatelessWidget {
-  final VoidCallback onPressed;
-  final String title;
-
-  const CustomTextButton({
-    required this.onPressed,
-    required this.title,
-    Key? key,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return TextButton(
-      style: TextButton.styleFrom(
-        padding: const EdgeInsets.symmetric(horizontal: 30, vertical: 20),
-        backgroundColor: Theme.of(context).accentColor.withOpacity(0.3),
-        shape: const RoundedRectangleBorder(
-          borderRadius: BorderRadius.all(
-            Radius.circular(8),
-          ),
-        ),
-      ),
-      onPressed: onPressed,
-      child: Text(
-        title,
-        style: Theme.of(context).textTheme.headline5,
-      ),
-    );
-  }
-}
